requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
streamlit==1.45.0
pypdf==5.1.0
python-dotenv==1.0.1
openai==1.78.0
faiss-cpu==1.9.0.post1
tiktoken==0.9.0
groq==0.28.0
sentence-transformers==3.0.1
numpy<2
python-docx
openpyxl
pandas
pytesseract
Pillow
gTTS
SpeechRecognition
pydub
streamlit-audiorecorder
fpdf2